chore: refactor to a separate subcommand
